Support removal of watchers from Zookeeper client.
Enables the application layer to stop watching over the channel
provided by the Watch APIs and use these APIs to clean up the
"connection.watchers" map. The channels that are no longer watched
upon will also be garbage collected with this change.

The expectation is that the application can use context with cancellation
along with the channel returned by the Watch APIs and call these
when the context is cancelled. Or the application can choose to call them
independently as well.

These functions can also be used to support the removeWatches API
that were added to Zookeeper 3.5.0 and above by actually making a
call to Zookeeper whenever needed.